Limit the size of the pagecache
This commit is based on the patch linked here: https://lwn.net/Articles/218890/
Original patch authored by Christoph Lameter <clameter@sgi.com>, and based on work by Aubrey Li <aubreylee@gmail.com>.
Minor changes included due to difference between reference kernel v2.6.20-rc5 and our kernel v2.6.33 source tree.

Patch/email comment text copied below:

----- BEGIN COPIED COMMENT TEXT -----

This is a patch using some of Aubrey's work plugging it in what is IMHO 
the right way. Feel free to improve on it. I have gotten repeatedly 
requests to be able to limit the pagecache. With the revised VM statistics 
this is now actually possile. I'd like to know more about possible uses of 
such a feature.

It may be useful to limit the size of the page cache for various reasons
such as

1. Insure that anonymous pages that may contain performance
   critical data is never subject to swap.

2. Insure rapid turnaround of pages in the cache.

3. Reserve memory for other uses? (Aubrey?)

We add a new variable "pagecache_ratio" to /proc/sys/vm/ that
defaults to 100 (all memory usable for the pagecache).

The size of the pagecache is the number of file backed
pages in a zone which is available through NR_FILE_PAGES.

We skip zones that contain too many page cache pages in
the page allocator which may cause us to enter reclaim.

If we enter reclaim and the number of page cache pages
is too high then we switch off swapping during reclaim
to avoid touching anonymous pages.

Signed-off-by: Christoph Lameter <clameter@sgi.com>

----- END COPIED COMMENT TEXT -----